telegram-bot/requirements.txt
aiogram
python-dotenv
aiohttp
asyncio
logging
typing
requests
urllib3